+++ NEWS.md
@@ -1,4 +1,12 @@
 
+# GiottoClass 0.3.3
+
+## bug fixes
+- fix flipping issue with `giottoAffineImage` for certain affine transforms
+
+## enhancements
+- `missing` method for `affine()` instantiates an `affine2d` object
+
 # GiottoClass 0.3.2 (2024/07/26)
 
 ## breaking changes